requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cycler==0.11.0
Django==3.2.7
djangorestframework==3.12.4
fonttools==4.28.4
idna==3.3
packaging==21.3
pandas==1.3.5
phonenumbers==8.12.41
Pillow==8.4.0
pycparser==2.21
pyfcm==1.5.1
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
twilio==7.5.0
urllib3==1.26.8
